Checkpoint: Associate names with tokens.
Also, go in function bodies when
looking for names. If we don't do this
we miss out on most of the names.

The current situation is not quite right,
because we need to know that the
"x" that was clicked was the one inside
"f", and not the current one.