Better support for working with multiple threads.
We had a bug, where things would get confused when we had a thread
open, and then we stepped.  The old thread would still appear to be
open, but it is actually "stale" and won't work.

Now we keep track of what threads are open properly, and updated them
as needed when the debugger executes.