Avoid doing STM on the fast-path.
This slightly reworks the logic for checking for external commands,
to avoid doing STM on the fast path.  This almost doubles the speed
of the debugger.

The basic heuristic is that there aren't very many external commands,
relative to the number of checks we do while running.  This is why
we want to be able to check quickly if there are external commnads or
not.  If not we keep going, if there are, we use the old logic we had.
The logic for this is implemnetd in Galua.Debuggar.CommandQueue