Hardcode using GCC for now.
If we don't do that, GHC will try to guess what the compiler is at runtime.
To do this, it forks the compiler and asks it for its version.
This, however, does not work in the context of Love, for example, as Love
installs a signal handler, and we miss the notification that a process
has ended, which in turn makes us complain that our child disappeared.

The same problem we have when running `luac`.